+++ src/utils.ts
@@ -1,2 +1,0 @@
-export const getRandomString = (): string =>
-    Math.random().toString(36).substr(2, 15)